Color Pallete/requirements.txt
colorgram.py
Pillow
pyperclip